pyproject.toml
[project]
name = "native-toolcall-and-mcp"
version = "0.1.0"
description = "Demo: Native Tool Calling vs MCP Tool Calling với LangGraph & Google Gemini"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.16",
    "langchain-core>=1.3.2",
    "langchain-google-genai>=4.2.2",
    "langchain-mcp-adapters>=0.2.2",
    "langgraph>=1.1.10",
    "mcp>=1.27.0",
    "python-dotenv>=1.2.2",
]
